requirements.txt
flask
PyGithub
google-generativeai